+++ package.json
@@ -23,7 +23,7 @@
 		"highlight.js": "^11.11.1",
 		"lowlight": "^3.3.0",
 		"lucide-react": "^0.544.0",
-		"next": "^15.6.0-canary.16",
+		"next": "^16.0.1",
 		"next-auth": "^5.0.0-beta.26",
 		"next-sitemap": "^4.2.3",
 		"prismjs": "^1.30.0",
