--- src/app/loading.tsx
@@ -1,5 +1,0 @@
-import PageContainer from '@/components/ui/containters/page-container'
-
-export default function Loading() {
-	return <PageContainer></PageContainer>
-}
